# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /polls:
    get:
      operationId: listOrganizerPolls
      tags: [polls]
      summary: List polls owned by the signed-in organizer
      responses:
        "200":
          description: Organizer poll history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OrganizerPollSummary"
        "401":
          description: Organizer authentication is required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createPoll
      tags: [polls]
      summary: Create an event date poll
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PollInput"
      responses:
        "201":
          description: Created poll
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreatedPoll"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Organizer authentication is required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Organizer creation limit exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /polls/{shareId}:
    get:
      operationId: getPoll
      tags: [polls]
      summary: Get a poll by its share ID
      parameters:
        - $ref: "#/components/parameters/ShareId"
      responses:
        "200":
          description: Poll details and responses
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Poll"
        "404":
          description: Poll not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updatePoll
      tags: [polls]
      summary: Update an owned poll and append candidate dates
      parameters:
        - $ref: "#/components/parameters/ShareId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PollUpdate"
      responses:
        "200":
          description: Updated poll
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Poll"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Organizer authentication is required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: The signed-in organizer does not own this poll
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Poll not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Poll is already closed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /polls/{shareId}/responses:
    post:
      operationId: createPollResponse
      tags: [polls]
      summary: Submit or update a participant response
      parameters:
        - $ref: "#/components/parameters/ShareId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PollResponseInput"
      responses:
        "201":
          description: Saved response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PollResponse"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Poll not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Response submission is closed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /polls/{shareId}/legacy-claim:
    post:
      operationId: claimLegacyPoll
      tags: [polls]
      summary: Claim a specific legacy poll with a one-time migration token
      parameters:
        - $ref: "#/components/parameters/ShareId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LegacyPollClaimInput"
      responses:
        "200":
          description: Legacy poll claimed by the signed-in organizer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Poll"
        "400":
          description: Invalid migration request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Organizer authentication is required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Invalid migration token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Legacy poll not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Legacy poll is already owned by another account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /polls/{shareId}/confirmation:
    post:
      operationId: confirmPoll
      tags: [polls]
      summary: Confirm the event date and close responses
      parameters:
        - $ref: "#/components/parameters/ShareId"
        - $ref: "#/components/parameters/AdminKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PollConfirmationInput"
      responses:
        "200":
          description: Confirmed poll
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Poll"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing management key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Invalid management key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Poll or candidate not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Poll is already closed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  parameters:
    ShareId:
      name: shareId
      in: path
      required: true
      schema:
        type: string
        minLength: 6
        maxLength: 32
    AdminKey:
      name: X-Admin-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 20
        maxLength: 128
  schemas:
...
    CalendarCandidateSlot:
      type: object
      required: [label, start, end]
      properties:
        label:
          type: string
        start:
          type: string
          format: date-time
        end:
          type: string
          format: date-time
...
    Poll:
      type: object
      required: [id, shareId, title, note, deadline, candidates, responses, createdAt, isClosed, confirmedCandidateId, canManage]
      properties:
        id:
          type: number
        shareId:
          type: string
        title:
          type: string
        note:
          type: string
        deadline:
          type: ["string", "null"]
          format: date-time
        candidates:
          type: array
          items:
            $ref: "#/components/schemas/PollCandidate"
        responses:
          type: array
          items:
            $ref: "#/components/schemas/PollResponse"
        createdAt:
          type: string
          format: date-time
        isClosed:
          type: boolean
        confirmedCandidateId:
          type: ["number", "null"]
        canManage:
          type: boolean
    CreatedPoll:
      allOf:
        - $ref: "#/components/schemas/Poll"
        - type: object
          required: [adminKey]
          properties:
            adminKey:
              type: string
              minLength: 20
              maxLength: 128
    OrganizerPollSummary:
      type: object
      required: [shareId, title, note, deadline, createdAt, isClosed, responseCount]
      properties:
        shareId:
          type: string
        title:
          type: string
        note:
          type: string
        deadline:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
        isClosed:
          type: boolean
        responseCount:
          type: number
    PollCandidate:
      type: object
      required: [id, label, start, end, yesCount, maybeCount, noCount]
      properties:
        id:
          type: number
        label:
          type: string
        start:
          type: ["string", "null"]
          format: date-time
        end:
          type: ["string", "null"]
          format: date-time
        yesCount:
          type: number
        maybeCount:
          type: number
        noCount:
          type: number
    PollResponse:
      type: object
      required: [id, name, answers, createdAt]
      properties:
        id:
          type: number
        name:
          type: string
        answers:
          type: array
          items:
            type: string
            enum: [yes, maybe, no]
        comment:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    PollInput:
      type: object
      required: [title, candidates]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 120
        note:
          type: string
          maxLength: 500
        deadline:
          type: ["string", "null"]
          format: date-time
        candidates:
          type: array
          minItems: 2
          maxItems: 20
          items:
            type: string
            minLength: 1
            maxLength: 100
        candidateSlots:
          type: array
          items:
            $ref: "#/components/schemas/CalendarCandidateSlot"
          maxItems: 20
    PollUpdate:
      type: object
      required: [title, note, newCandidates]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 100
        note:
          type: string
          maxLength: 500
        newCandidates:
          type: array
          maxItems: 20
          items:
            type: string
            minLength: 1
            maxLength: 100
    PollResponseInput:
      type: object
      required: [name, answers]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 80
        answers:
          type: array
          minItems: 1
          maxItems: 20
          items:
            type: string
            enum: [yes, maybe, no]
        comment:
          type: ["string", "null"]
          maxLength: 500
    PollConfirmationInput:
      type: object
      required: [candidateId]
      properties:
        candidateId:
          type: number
    LegacyPollClaimInput:
      type: object
      required: [claimToken]
      properties:
        claimToken:
          type: string
          minLength: 20
          maxLength: 128
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
        code:
          type: string
          enum: [google_calendar_reconnect_required]