# lib/api-zod/src/generated/types/legacyPollClaimInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface LegacyPollClaimInput {
  /**
     * @minLength 20
     * @maxLength 128
     */
  claimToken: string;
}