# lib/api-zod/src/generated/types/poll.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PollCandidate } from './pollCandidate';
import type { PollResponse } from './pollResponse';

export interface Poll {
  id: number;
  shareId: string;
  title: string;
  note: string;
  /** @nullable */
  deadline: Date | null;
  candidates: PollCandidate[];
  responses: PollResponse[];
  createdAt: Date;
  isClosed: boolean;
  /** @nullable */
  confirmedCandidateId: number | null;
  canManage: boolean;
}