# lib/api-zod/src/generated/types/pollCandidate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PollCandidate {
  id: number;
  label: string;
  /** @nullable */
  start: Date | null;
  /** @nullable */
  end: Date | null;
  yesCount: number;
  maybeCount: number;
  noCount: number;
}