# lib/api-zod/src/generated/types/pollInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CalendarCandidateSlot } from './calendarCandidateSlot';

export interface PollInput {
  /**
     * @minLength 1
     * @maxLength 120
     */
  title: string;
  /** @maxLength 500 */
  note?: string;
  /** @nullable */
  deadline?: Date | null;
  /**
     * @minItems 2
     * @maxItems 20
     * @items.minLength 1
     * @items.maxLength 100
     */
  candidates: string[];
  /** @maxItems 20 */
  candidateSlots?: CalendarCandidateSlot[];
}