# lib/api-zod/src/generated/types/pollResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PollResponseAnswersItem } from './pollResponseAnswersItem';

export interface PollResponse {
  id: number;
  name: string;
  answers: PollResponseAnswersItem[];
  /** @nullable */
  comment?: string | null;
  createdAt: Date;
}